backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
astroid==3.0.3
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
axios==0.4.0
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.11.1
black==24.1.1
bleach==6.2.0
blinker==1.7.0
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
chromedriver==2.24.1
click==8.1.7
cloudpathlib==0.16.0
comm==0.2.1
confection==0.1.4
constantly==23.10.4
contourpy==1.2.0
cryptography==42.0.1
cycler==0.12.1
cymem==2.0.8
Cython==0.29.37
daphne==4.0.0
debugpy==1.8.1
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
dnspython==2.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
executing==2.0.1
Faker==24.1.0
fastapi==0.104.1
fastcluster==1.2.6
fastjsonschema==2.20.0
ffmpeg==1.4
filelock==3.13.1
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.42
google==3.0.0
google-api-core==2.16.2
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
h11==0.14.0
hdbscan==0.8.33
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.20.3
hyperlink==21.0.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
incremental==22.10.0
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.22.1
ipywidgets==8.1.5
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.28
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
langcodes==3.3.0
llvmlite==0.42.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.4
notebook==7.2.2
notebook_shim==0.2.4
numba==0.59.0
numpy==1.26.3
oauthlib==3.2.2
openai==1.2.4
opencv-python==4.9.0.80
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.18.0
pluggy==1.5.0
preshed==3.0.9
proglog==0.1.10
prometheus_client==0.21.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
pygame==2.5.2
Pygments==2.17.2
PyJWT==1.6.4
pylint==3.0.3
pymongo==4.10.1
pynndescent==0.5.11
pyOpenSSL==24.0.0
pyparsing==3.1.1
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.2.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
selenium==4.16.0
Send2Trash==1.8.3
service-identity==24.1.0
setuptools==69.0.3
shapely==2.0.4
singledispatch==4.1.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.27
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
streamlit==1.31.1
sympy==1.12
tenacity==8.2.3
terminado==0.18.1
thinc==8.2.3
threadpoolctl==3.3.0
tiktoken==0.6.0
tinycss2==1.4.0
tokenizers==0.15.1
toml==0.10.2
tomlkit==0.12.3
toolz==0.12.1
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.37.2
trio==0.24.0
trio-websocket==0.11.1
Twisted==23.10.0
txaio==23.1.1
typer==0.9.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
umap-learn==0.5.5
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.27.0
validators==0.22.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webcolors==24.11.1
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.42.0
widgetsnbextension==4.0.13
wsproto==1.2.0
yarl==1.9.4
youtube-transcript-api==0.6.2
zipp==3.17.0
zope.interface==6.1
